cx_Freeze

+++ Cloning-Stuff/setHostname/setHostname.py
@@ -26,8 +26,6 @@ import logging
 from cryptography.fernet import Fernet
 from libs.LoggerConfiguration import configure_logging
 from libs.Worker import Worker
-import signal
-import sys
 import atexit
 import shutil
 
@@ -49,13 +47,12 @@ class setHostname():
         self.keyFile = os.path.join(self.rootDir, 'libs', 'key.key')
         self.tmpPath = os.path.join(self.rootDir, 'tmp/')
         self.logger = logging.getLogger('setHostname')
-        
+
         # catch terminating Signal
         atexit.register(self.exit_handler)
 
         self.read_cli_args()
         self.config = self._load_yml()
-        
 
     def _load_yml(self):
         """ Load the yaml file config.yaml """
@@ -122,18 +119,20 @@ class setHostname():
         encMessage = fernet.encrypt(self.estring.encode())
         print("%s: %s" % (self.estring, encMessage.decode()))
         print("\nUse this hash in your config File for sensible data, e.g. passwords")
-        
+
     def exit_handler(self):
         """ do something on sys.exit() """
         # be sure to delete tmp/ dir
-        try:
-            shutil.rmtree(self.tmpPath)
-        except OSError as e:
-            print("Error: %s : %s" % (self.tmpPath, e.strerror))
+        if (os.path.exists(self.tmpPath) is True):
+            try:
+                shutil.rmtree(self.tmpPath)
+            except OSError as e:
+                print("Error: %s : %s" % (self.tmpPath, e.strerror))
 
 
 if __name__ == "__main__":
     # load logging Config
     configure_logging()
     sethostname = setHostname()
-    sethostname.start()
+    # sethostname.start()
+    print("Juhuuuuuu")
